rework SyncProxy to use globally available channels
the SyncProxy is now a globally available OnceCell that transparently
offers a struct with methods to send syncmessages and rpccalls.
This allows me to get rid of the globally passed around syncproxy var.

The Syncable trait provides default implementations that access this
global so the library user usually does not need to use it directly.

Further the Syncable trait now provides default implementations for it's
methods and only a const for the CLASS name must be set.

It isn't very idiomatic rust and takes away some freedom from the
library user but i think that it's an overall nicer solution and
simplifies the code big time.